add module for dev-only utils
